requirements.txt
asgiref==3.8.1
boto3==1.34.89
botocore==1.34.89
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
dj-database-url==2.1.0
Django==4.2.11
django-admin-inline-paginator==0.4.0
django-admin-rangefilter==0.12.1
django-adminfilters==2.4.3
django-cors-headers==4.3.1
django-countries==7.6.1
django-debug-toolbar==4.3.0
django-filter==23.5
django-rest-knox==4.2.0
django-storages==1.14.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
awscli>=1.32.38,<2
django-storages>=1.14.2,<2
drf-yasg==1.21.7
Faker==24.11.0
flake8==7.0.0
gunicorn==22.0.0
idna==3.6
inflection==0.5.1
install==1.3.5
jmespath==1.0.1
mccabe==0.7.0
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==1.26.18
whitenoise==6.6.0
